LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-005-expected.txt


FAIL Target should not scroll with viewport when anchor is in fixed-positioned scroller assert_equals: expected 100 but got 500
